initial development
Restructured the main program, refined the main field datastructure, and added some data structure error checking. Also updated the main program to use the "n" live data from Speeduino, rather than the legacy "A", and slowed it down to a 10Hz rate (it was recording at 62Hz running naturally). Not entirely sure what the optimal rate is, but 10Hz is a good place to start, I suspect.